Initial commit for Stubby / Processing interface
The basic layout for a Processing library is included here, with an ant
script which will build everything in the correct locations.  The
following features are supported so far:
-Build the .zip file with 'ant' default task; unzip the resulting
stubby.zip into "<Processing Sketch Root>/libraries" to show the new
library under the "Sketch -> Import Library -> Stubby Control Interface"
menu.
-When making a sketch including Stubby control, you need to include the
serial library as well
-You can create a new Stubby instance using either the automatic serial
port discovery (it connects to the first serial port containing 'stubby'
in the name) or a specified port
-Some of the methods are included (but not yet implemented), such as
moveForward(), rotate(), etc.