Massive refactoring; see below for details.
Just finished a massive refactoring to prepare for Processing
integration.  Some high level changes include:
-Adding sub folders for common files (i.e. 'controllers' folder contains
universal_controller and processing drivers; 'util' contains trig and
convert, etc).
-Move all the controller-specific code to the driver files
(universal_controller and processing).  The main loop checks to see
which controller has registered itself (via the
MESSAGE_ANNOUNCE_CONTROL_ID message), and passes control to the
appropriate executor function.
-Extracted the actual movement functions (the ones which call the given
gait and move the legs), so that the drivers can call them.
-Removed DEBUG_SIMULATION defines / code.  This simulation code is
mostly obselete now anyway, as I used it mostly to verify IK values.  If
we need it again in the future, we can always retrieve from git.

As of the time of this commit, the code has been uploaded (both to
Stubby and Universal Controller) and appears to be working.  Be sure to
apply the updates to both Universal Controller as well as Stubby; there
are some changes in Stubby which require a relatively frequent
MESSAGE_ANNOUNCE_CONTROL_ID message to be sent, and until tonight the UC
did not send this nearly enough; currently it sends it once every 2 - 3
seconds.