Simplify move and turn commands
Previously, move and turn had a bit of a fuzzy definition with a fait
bit of functional overlap (you could turn in place by moving without
linear velocity but with rotational velocity).  This is now separated,
and much more clearly defined: to move in a direction, use the move
command; to turn in place, use the turn command.  The executor code has
also been cleaned up quite a bit (still a WIP, though).