Added the concept of 'heading verification'
This happens at the beginning of each heading-dependent operation (move,
turn, etc).  We repeatedly poll the magnetometer until we have a stable
reading.  We then use that reading as the set point for the following
operation.  This appears to work very well; for a short (maybe 0.5
second) pause before each action, the accuracy of the action goes way
up.